Added markdown to notes; fixed date cross over bug.
You can now use markdown in the notes field input. This will be styled in the
tabular output.

A long-term bug existed where we could not enter time that started on one date
and ended on another. Thus if work was done from 23:00 to 01:00 we now get the
correct number of hours, 2.

Committer: Joseph Hallenbeck <joseph@joehallenbeck.com>