Docker and convert database to SQLITE
The following commit converts the database back end from MYSQL to an
SQLITE database stored locally within the repository. This makes backing
up and maintaining the TimeKeeper application easier as it is a
self-contained unit. Additionally, adds support for installation via
Docker and begins the process of converting the application to Slim with
support for PHP 8.0