Host Slim API
Update the Slim skeleton and base docker-compose to route requests on
port 8081 to `api/public/index.php`. Update unit tests on the skeleton
to use the `PhrophecyTrait`. And update xdebug settings for Xdebug 3.

Confirm both unit tests work inside docker container and requests to
localhost:8081 route to the "hello world" page in the slim container.