Implement JSON-API styled error responses
The commit implements JONS-API error responses in the form of a
collection of Error Objects. This allows for multiple validation errors
to occur in a single request as well as both an overall HTTP status code
response and seperate specific errors with their own codes.

Added stubs for ListTimeAccountAction and LisTimeTaskAction.