Move Accounts into Module, Added Tasks
Move TimeAccount and related classes into it's own Account module and
refactored the Account related logic out of TimeService and
TimeRepository so that Account has it's own Service and Repository
classes.

Added a ViewAccountAction to request a single account record.

Added the Tasks module which mirrors Accounts along with a
ListTaskAction and ViewTaskAction.